HACK: Make parceljs ignore the webpack-only require
The `if` surrounding the require of `!raw-loader!./js.py` should be enough to limit the processing of this line only by webpack. However, looks like parceljs is greedy and tries to follow such path. Then it fails for a code path that will never be executed by him.

The code change make parceljs ignore such require. Is not beautiful and should be replaced when parceljs provides something nicer.